+++ .changeset/nervous-crews-yawn.md
@@ -1,0 +1,5 @@
+---
+"@opennextjs/cloudflare": patch
+---
+
+Use kebab-case for the KV Cache.
