+++ packages/cloudflare/templates/defaults/open-next.config.ts
@@ -1,6 +1,6 @@
 // default open-next.config.ts file created by @opennextjs/cloudflare
 
-import cache from "@opennextjs/cloudflare/kvCache";
+import cache from "@opennextjs/cloudflare/kv-cache";
 
 const config = {
   default: {
